fix: Fix crash with latest Shogi

+++ common/src/main/java/net/blay09/mods/waystones/util/ExperienceUtils.java
@@ -20,7 +20,7 @@ public class ExperienceUtils {
         return level;
     }
 
-    public static int calculateDisplayedLevelCostFromExperiencePoints(int currentLevel, int availableXp, int requiredXp) {
+    public static int calculateDisplayedLevelCostFromExperiencePoints(int currentLevel, long availableXp, int requiredXp) {
         if (requiredXp <= 0) {
             return 0;
         }
